+++ src/test/java/com/moabam/api/domain/item/repository/InventorySearchRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.moabam.api.domain.repository;
+package com.moabam.api.domain.item.repository;
 
 import static com.moabam.support.fixture.InventoryFixture.*;
 import static com.moabam.support.fixture.ItemFixture.*;
@@ -16,9 +16,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.moabam.api.domain.item.Inventory;
 import com.moabam.api.domain.item.Item;
 import com.moabam.api.domain.item.ItemType;
-import com.moabam.api.domain.item.repository.InventoryRepository;
-import com.moabam.api.domain.item.repository.InventorySearchRepository;
-import com.moabam.api.domain.item.repository.ItemRepository;
 import com.moabam.api.domain.member.Member;
 import com.moabam.api.domain.member.repository.MemberRepository;
 import com.moabam.support.annotation.QuerydslRepositoryTest;
