+++ src/test/java/com/moabam/api/domain/product/ProductTest.java
@@ -1,11 +1,10 @@
-package com.moabam.api.domain.entity;
+package com.moabam.api.domain.product;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.moabam.api.domain.product.Product;
 import com.moabam.global.error.exception.BadRequestException;
 
 class ProductTest {
